+++ CHANGELOG.md
@@ -19,6 +19,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - activityWrapper for handling different activities
 - textActivity for reading info texts at stations
 - loading Activities for Stations via api call
+- MultiChoiceActivity view
+- handle choose and reveal answers of multi choice view
 
 ## [0.0.0] - Date
 ### Added